[refactor] #60 network 모듈이 앱의 기획과 너무 의존되어 있어 그 기획에 대한 책임을 data 모듈에 넘기고 통신을 위한 도구의 역할을 하도록 재구성
