+++ core/data/src/main/java/com/example/data/datasource/remote/DummyRemoteDataSource.kt
@@ -1,8 +1,8 @@
 package com.example.data.datasource.remote
 
-import com.example.network.response.BaseResponse
-import com.example.network.response.DummyResponse
-import com.example.network.service.DummyService
+import com.example.data.model.response.BaseResponse
+import com.example.data.model.response.DummyResponse
+import com.example.data.service.DummyService
 import kotlinx.serialization.InternalSerializationApi
 import javax.inject.Inject
