+++ core/data/src/main/java/com/example/data/model/response/dummyResponse.kt
@@ -1,0 +1,9 @@
+package com.example.data.model.response
+
+import com.example.domain.model.Dummy
+
+data class DummyResponse(
+    val dummyName: String,
+){
+    fun toDummyEntity(): Dummy = Dummy(dummyName = dummyName)
+}
