+++ core/data/src/main/java/com/example/data/service/DummyService.kt
@@ -1,15 +1,15 @@
-package com.example.network.service
+package com.example.data.service
 
-import com.example.network.NetworkConstants
-import com.example.network.response.BaseResponse
-import com.example.network.response.DummyResponse
+import com.example.data.ApiConstants
+import com.example.data.model.response.BaseResponse
+import com.example.data.model.response.DummyResponse
 import kotlinx.serialization.InternalSerializationApi
 import retrofit2.http.GET
 import retrofit2.http.Path
 
 @OptIn(InternalSerializationApi::class)
 interface DummyService {
-    @GET("${NetworkConstants.API}/${NetworkConstants.VERSIONS}/dummy/{dummyId}")
+    @GET("${ApiConstants.API}/${ApiConstants.VERSIONS}/dummy/{dummyId}")
     suspend fun getDummy(
         @Path("dummyId") dummyId: Long,
     ): BaseResponse<DummyResponse>
