--- core/network/src/main/java/com/example/network/response/dummyResponse.kt
@@ -1,5 +1,0 @@
-package com.example.network.response
-
-data class DummyResponse(
-    val dummyName: String,
-)
